Fixed maxPrice and minPrice logic, added price filter output
