+++ src/app/modules/core/advertisement/components/advertisement-list/sidebar/sidebar-container/sidebar-container.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core'
+import { Advertisement } from 'src/app/shared/models/advertisement'
+
+import { AdvertisementService } from './../../../../advertisement.service'
 
 @Component({
   selector: 'cav-sidebar-container',
@@ -6,7 +9,39 @@ import { Component, OnInit } from '@angular/core'
   styleUrls: ['./sidebar-container.component.scss'],
 })
 export class SidebarContainerComponent implements OnInit {
-  constructor() {}
+  constructor(private advertisementService: AdvertisementService) {}
 
-  ngOnInit(): void {}
+  maxPriceAdvertisement: Advertisement[]
+  minPriceAdvertisement: Advertisement[]
+  maxPrice: number
+  minPrice: number
+
+  ngOnInit(): void {
+    this.maxPrice = this.returnMaxPrice()
+    this.minPrice = this.returnMinPrice()
+  }
+
+  returnMaxPrice(): number {
+    this.advertisementService
+      .returnMaxPriceProperty()
+      .subscribe(
+        (maxPriceAdvertisement) => (this.maxPriceAdvertisement = maxPriceAdvertisement)
+      )
+    const maxPricestring = this.maxPriceAdvertisement.map((x) => x.price).toString()
+    return parseInt(maxPricestring, 0)
+  }
+
+  returnMinPrice(): number {
+    this.advertisementService
+      .returnMinPriceProperty()
+      .subscribe(
+        (minPriceAdvertisement) => (this.minPriceAdvertisement = minPriceAdvertisement)
+      )
+    const minPricestring = this.minPriceAdvertisement.map((x) => x.price).toString()
+    return parseInt(minPricestring, 0)
+  }
+
+  applyPriceFilter(filter: number): void {
+    console.log(`Value received from slider: ${filter}`)
+  }
 }
